+++ src/resources/js/blitzInjectScript.js
@@ -1,65 +1,110 @@
-var Blitz = (function () {
-    function Blitz() {
-        var _this = this;
-        this.processed = 0;
-        if (!document.dispatchEvent(new CustomEvent('beforeBlitzInjectAll', { cancelable: true }))) {
-            return;
-        }
-        var urls = {};
-        var elements = document.querySelectorAll('.blitz-inject:not(.blitz-inject--injected)');
-        this.elementsToProcess = elements.length;
-        elements.forEach(function (element) {
-            var _a;
-            var priority = element.getAttribute('data-blitz-priority');
-            var url = {
-                id: element.getAttribute('data-blitz-id'),
-                uri: element.getAttribute('data-blitz-uri'),
-                params: element.getAttribute('data-blitz-params'),
-                element: null,
-                response: null,
-            };
-            if (!document.dispatchEvent(new CustomEvent('beforeBlitzInject', { cancelable: true, detail: url }))) {
-                return;
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
             }
-            var key = url.uri + (url.params && '?' + url.params);
-            urls[key] = (_a = urls[key]) !== null && _a !== void 0 ? _a : [];
-            urls[key].push(url);
-        });
-        var _loop_1 = function (key) {
-            fetch(key)
-                .then(function (response) {
-                if (response.status >= 200 && response.status < 300) {
-                    return response.text();
-                }
-            })
-                .then(function (response) {
-                _this.replaceUrls(urls[key], response);
-            });
-        };
-        for (var key in urls) {
-            _loop_1(key);
-        }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
     }
-    Blitz.prototype.replaceUrls = function (urls, response) {
-        var _this = this;
-        urls.forEach(function (url) {
-            var element = document.getElementById('blitz-inject-' + url.id);
-            if (element) {
-                url.element = element;
-                element.innerHTML = response;
-                element.classList.add('blitz-inject--injected');
+};
+document.addEventListener('{injectScriptEvent}', injectElements);
+function injectElements() {
+    return __awaiter(this, void 0, void 0, function () {
+        var elements, injectElements, promises, url;
+        return __generator(this, function (_a) {
+            switch (_a.label) {
+                case 0:
+                    if (!document.dispatchEvent(new CustomEvent('beforeBlitzInjectAll', { cancelable: true }))) {
+                        return [2];
+                    }
+                    elements = document.querySelectorAll('.blitz-inject:not(.blitz-inject--injected)');
+                    injectElements = {};
+                    promises = [];
+                    elements.forEach(function (element) {
+                        var _a;
+                        var priority = element.getAttribute('data-blitz-priority');
+                        var injectElement = {
+                            element: element,
+                            id: element.getAttribute('data-blitz-id'),
+                            uri: element.getAttribute('data-blitz-uri'),
+                            params: element.getAttribute('data-blitz-params'),
+                            property: element.getAttribute('data-blitz-property'),
+                        };
+                        if (document.dispatchEvent(new CustomEvent('beforeBlitzInject', { cancelable: true, detail: injectElement }))) {
+                            var url = injectElement.uri + (injectElement.params && '?' + injectElement.params);
+                            injectElements[url] = (_a = injectElements[url]) !== null && _a !== void 0 ? _a : [];
+                            injectElements[url].push(injectElement);
+                        }
+                    });
+                    for (url in injectElements) {
+                        promises.push(replaceUrls(url, injectElements[url]));
+                    }
+                    return [4, Promise.all(promises)];
+                case 1:
+                    _a.sent();
+                    document.dispatchEvent(new CustomEvent('afterBlitzInjectAll'));
+                    return [2];
             }
-            url.response = response;
-            document.dispatchEvent(new CustomEvent('afterBlitzInject', { detail: url }));
-            _this.processed++;
         });
-        if (this.processed >= this.elementsToProcess) {
-            document.dispatchEvent(new CustomEvent('afterBlitzInjectAll'));
-        }
-    };
-    return Blitz;
-}());
-;
-document.addEventListener('{injectScriptEvent}', function () {
-    new Blitz();
-});
+    });
+}
+function replaceUrls(url, injectElements) {
+    return __awaiter(this, void 0, void 0, function () {
+        var response, responseText, responseJson;
+        return __generator(this, function (_a) {
+            switch (_a.label) {
+                case 0: return [4, fetch(url)];
+                case 1:
+                    response = _a.sent();
+                    if (response.status >= 300) {
+                        return [2, null];
+                    }
+                    return [4, response.text()];
+                case 2:
+                    responseText = _a.sent();
+                    if (url.indexOf('blitz/csrf/json') !== -1) {
+                        responseJson = JSON.parse(responseText);
+                    }
+                    injectElements.forEach(function (injectElement) {
+                        var _a;
+                        if (injectElement.property) {
+                            injectElement.element.innerHTML = (_a = responseJson[injectElement.property]) !== null && _a !== void 0 ? _a : '';
+                        }
+                        else {
+                            injectElement.element.innerHTML = responseText;
+                        }
+                        injectElement.element.classList.add('blitz-inject--injected');
+                        document.dispatchEvent(new CustomEvent('afterBlitzInject', { detail: injectElement }));
+                    });
+                    return [2];
+            }
+        });
+    });
+}