+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.11.1 - Unreleased
 ### Changed
+- The dynamic content inject script now only makes a single AJAX request per unique URL, including CSRF inputs and tokens ([#329](https://github.com/putyourlightson/craft-blitz/issues/329)).
 - The refresh cache job now ignores exceptions when executing element queries ([#355](https://github.com/putyourlightson/craft-blitz/issues/355)).
 
 ## 3.11.0 - 2021-11-15